refactor: split CLIcore_script_builtin.c into focused sub-modules
CLIcore_script_builtin.c was a 2,622-line file mixing four
unrelated concerns: event traps, I/O commands, system
inspection, and a defer stack.

Split into four files:

  CLIcore_script_traps.c       (493 lines)
    Signal-based trap table + non-blocking engine
    event trap system (STREAM:/FPS:/PROC: prefixes)
    with lazy-connect, throttle, and fire-count limit.
    Functions: cli_trap_run, cli_trap_run_exit,
    etrap_connect, etrap_check_fire,
    cli_engine_traps_poll, cli_engine_traps_cleanup

  CLIcore_script_cmd_io.c      (746 lines)
    Interactive I/O and variable commands.
    Functions: cli_fps_set_param, cli_cmd_fpsset,
    cli_cmd_echo, cli_cmd_read, cli_cmd_export,
    cli_cmd_shift, cli_cmd_printf

  CLIcore_script_cmd_inspect.c (1,282 lines)
    System inspection with table and JSON output.
    Functions: json_escape_str, emit_fps_json_body,
    cli_cmd_fpslist, cli_cmd_fpsdump,
    emit_streams_json_body, cli_cmd_streamlist,
    emit_procs_json_body, cli_cmd_proclist,
    cli_cmd_milkquery

  CLIcore_script_cmd_defer.c   (139 lines)
    LIFO cleanup stack.
    Functions: cli_cmd_defer, cli_defer_run

CLIcore_script_builtin.c retained as a stub
explaining the split (compiles to zero symbols).

All 4 new files added to MILKSCRIPT_SOURCES in
CMakeLists.txt. All public symbols remain declared
in CLIcore_script.h — no other file needs updating.

Verified: 276/276 CLI robustness tests pass.
